odbcBackend: refactor, SQLite sample db
- Refactor the code for the ODBC backend plugin into multiple modules
- Include a sample DB (SQLite file + .sql file)
- Remove commented out tests (copied from the backend plugin for files)
