+++ src/plugins/backend_odbc/CMakeLists.txt
@@ -10,7 +10,7 @@ endif ()
 
 add_plugin (
 	backend_odbc
-	SOURCES backend_odbc.h backend_odbc.c backendprivate_odbc.h helpers_odbc.c
+	SOURCES backend_odbc.h backend_odbc.c backend_odbc_helpers.h backend_odbc_helpers.c backend_odbc_general.h backend_odbc_general.c backend_odbc_get.h backend_odbc_get.c
 	INCLUDE_DIRECTORIES ${ODBC_INCLUDE_DIRS}
 	LINK_LIBRARIES ${ODBC_LIBRARIES}
 	ADD_TEST COMPONENT libelektra${SO_VERSION})
