+++ src/plugins/backend_odbc/backend_odbc.h
@@ -15,9 +15,10 @@ int ELEKTRA_PLUGIN_FUNCTION (open) (Plugin * handle, Key * errorKey);
 int ELEKTRA_PLUGIN_FUNCTION (init) (Plugin * handle, KeySet * definition, Key * parentKey);
 int ELEKTRA_PLUGIN_FUNCTION (get) (Plugin * handle, KeySet * ks, Key * parentKey);
 int ELEKTRA_PLUGIN_FUNCTION (set) (Plugin * handle, KeySet * ks, Key * parentKey);
+int ELEKTRA_PLUGIN_FUNCTION (commit) (Plugin * plugin, KeySet * ks, Key * parentKey);
 int ELEKTRA_PLUGIN_FUNCTION (error) (Plugin * handle, KeySet * ks, Key * parentKey);
 int ELEKTRA_PLUGIN_FUNCTION (close) (Plugin * handle, Key * errorKey);
 
 Plugin * ELEKTRA_PLUGIN_EXPORT;
 
-#endif // ELEKTRA_BACKEND_ODBC_SH
+#endif // ELEKTRA_BACKEND_ODBC_H
