+++ src/tools/kdb/mountOdbc.cpp
@@ -24,8 +24,8 @@ int MountOdbcCommand::execute (Cmdline const & cl)
 	kdb::Key keyBackend ("system:/elektra/mountpoints/user:\\/odbc/plugins/backend", KEY_END);
 	kdb::Key keyBackendName ("system:/elektra/mountpoints/user:\\/odbc/plugins/backend/name", KEY_END);
 	kdb::Key keyDataSourceName ("system:/elektra/mountpoints/user:\\/odbc/definition/dataSourceName", KEY_END);
-	kdb::Key keyUserName ("system:/elektra/mountpoints/user:\\/odbc/definition/userName", KEY_END);
-	kdb::Key keyPassword ("system:/elektra/mountpoints/user:\\/odbc/definition/password", KEY_END);
+	// kdb::Key keyUserName ("system:/elektra/mountpoints/user:\\/odbc/definition/userName", KEY_END);
+	// kdb::Key keyPassword ("system:/elektra/mountpoints/user:\\/odbc/definition/password", KEY_END);
 	kdb::Key keyTableName ("system:/elektra/mountpoints/user:\\/odbc/definition/table/name", KEY_END);
 	kdb::Key keyKeyColName ("system:/elektra/mountpoints/user:\\/odbc/definition/table/keyColName", KEY_END);
 	kdb::Key keyValColName ("system:/elektra/mountpoints/user:\\/odbc/definition/table/valColName", KEY_END);
@@ -34,26 +34,26 @@ int MountOdbcCommand::execute (Cmdline const & cl)
 	kdb::Key keyMetaTableMetaKeyColName ("system:/elektra/mountpoints/user:\\/odbc/definition/metaTable/metaKeyColName", KEY_END);
 	kdb::Key keyMetaTableMetaValColName ("system:/elektra/mountpoints/user:\\/odbc/definition/metaTable/metaValColName", KEY_END);
 
+	/* SQlite */
 	keyBackendName.setString ("backend_odbc");
-	keyDataSourceName.setString ("Pelektra");
-	keyUserName.setString ("flo");
-	keyPassword.setString ("elektra");
-	keyTableName.setString ("elektra");
-	keyKeyColName.setString ("key");
-	keyValColName.setString ("val");
-	keyMetaTableName.setString ("elektrameta");
-	keyMetaTableKeyColName.setString ("key");
-	keyMetaTableMetaKeyColName.setString ("metakey");
-	keyMetaTableMetaValColName.setString ("metaval");
-
+	keyDataSourceName.setString ("Selektra");
+	// keyUserName.setString ("flo");
+	// keyPassword.setString ("elektra");
+	keyTableName.setString ("elektraKeys");
+	keyKeyColName.setString ("keyName");
+	keyValColName.setString ("keyValue");
+	keyMetaTableName.setString ("metaKeys");
+	keyMetaTableKeyColName.setString ("keyName");
+	keyMetaTableMetaKeyColName.setString ("metaKeyName");
+	keyMetaTableMetaValColName.setString ("metaKeyValue");
 
 	kdb::KeySet ksOdbcConfig;
 	ksOdbcConfig.append (keyMp);
 	ksOdbcConfig.append (keyBackend);
 	ksOdbcConfig.append (keyBackendName);
 	ksOdbcConfig.append (keyDataSourceName);
-	ksOdbcConfig.append (keyUserName);
-	ksOdbcConfig.append (keyPassword);
+	// ksOdbcConfig.append (keyUserName);
+	// ksOdbcConfig.append (keyPassword);
 	ksOdbcConfig.append (keyTableName);
 	ksOdbcConfig.append (keyKeyColName);
 	ksOdbcConfig.append (keyValColName);
